refactor: migrate hooks (use-editor-flagging) from web/app/ce to web/app/core
